Add custom pin function option
