+++ src/main.ts
@@ -1,3 +1,3 @@
 export { prepare } from "./prepare.js";
 export { resolve } from "./resolve.js";
-export { optionalSetup } from "./ipfs-http-client-setup.js";
+export { optionalSetup } from "./setup.js";
