+++ src/prepare.ts
@@ -1,4 +1,4 @@
-import { globalIpfsClient } from "./ipfs-http-client-setup.js";
+import { globalPinFunction } from "./setup.js";
 
 function zip(firstArray: string[], secondArray: string[]) {
   let str = "";
@@ -13,54 +13,7 @@ function zip(firstArray: string[], secondArray: string[]) {
   return str;
 }
 
-type FileContent = Uint8Array;
-// ipfs.add all supports these:
-// May add support later.
-// | String
-// | Iterable<Uint8Array>
-// | AsyncIterable<Uint8Array>
-// | ReadableStream<Uint8Array>;
 
-async function sendToPinned(content: FileContent, _name = null) {
-  const ipfs = await globalIpfsClient;
-
-  const pinningServices = await ipfs.pin.remote.service.ls();
-
-  const hasRemotePinningService = pinningServices.length > 0;
-
-  if (!hasRemotePinningService) {
-    const noPinningService = "No remote pinning service connected.";
-    throw noPinningService;
-  } else {
-    let path = "";
-    if (_name) {
-      path = `/${_name}`;
-    }
-
-    const res = await ipfs.add(    
-      { content, path },
-      { wrapWithDirectory: path !== "" }
-    );
-    const { cid } = res;
-    await ipfs.pin.remote
-      .add(cid, {
-        service: pinningServices[0].service,
-        name: "Tableland Upload",
-      })
-      .catch((err: any) => {
-        const message: string = err.message;
-        if (message.includes("DUPLICATE_OBJECT")) {
-          console.log(
-            "Good news; that CID is already pinned to your pinning service."
-          );
-          return;
-        }
-        throw err;
-      });
-
-    return cid.toV1().toString();
-  }
-}
 
 // If you want to save a string to IPFS, simply encode it
 // Do this using `new TextEncoder().encode("Your string here")`
@@ -71,7 +24,7 @@ async function prepare(strings: TemplateStringsArray, ...values: any[]) {
     let res = value;
     switch (true) {
       case value instanceof Uint8Array:
-        res = await sendToPinned(value);
+        res = await globalPinFunction(value, null);
         break;
     }
     return res;
